+++ plugins/web_adapter/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 
-*None.*
+- Fix inaccurate timeout type detection using `xhr.readyState` instead of timer existence.
 
 ## 2.1.1